backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
openai>=1.30.0
pandas>=2.2.0
httpx>=0.27.0      # for TestClient in tests
pytest>=8.2.0
pytest-asyncio>=0.23.0
ruff>=0.4.0
